noa guest+DBcontroller

+++ src/system/DBController.java
@@ -20,20 +20,6 @@ public class DBController {
 
 
     /***************************************Guest function******************************************/
-    /**
-     * this function check if the guest is the one who call the function. and then
-     * @param guest
-     * @param fan
-     * @throws AlreadyExistException
-     */
-    public void addFan(Role guest , Fan fan) throws AlreadyExistException {db.addFan(fan);}
-
-    public boolean existMember(String id) {
-        return db.existMember(id);
-    }
-    public Role getMember(String id) throws MemberNotExist {
-        return db.getMember(id) ;
-    }
 
 
     /***************************************Getters******************************************/
@@ -49,59 +35,90 @@ public class DBController {
         return db.getLeagues();
     }
 
+    public Role getMember(String id) throws MemberNotExist {
+        if(! db.existMember(id))
+            throw new MemberNotExist();
+        return db.getMember(id) ;
+    }
     /***************************************delete function function******************************************/
 
-    public void deleteRole(String id) {
+    public void deleteRole(String id) throws MemberNotExist{
+        if(! db.existMember(id))
+            throw new MemberNotExist();
         db.removeRole(id);
     }
 
     /***************************************add function******************************************/
-    public void addSeason(Season season) {
+    public void addFan(Fan fan) throws AlreadyExistException {
+        if(db.existMember(fan.getUserMail()))
+            throw new AlreadyExistException();
+        db.addFan(fan);
+    }
+
+    public void addSeason(Season season) throws AlreadyExistException{
+        if(db.existSeason(season.getYear()))
+            throw new AlreadyExistException();
         db.addSeason(season);
     }
 
-    public void addLeague(League league) {
+    public void addLeague(League league) throws AlreadyExistException{
+        if(db.existLeague(league.getName()))
+            throw new AlreadyExistException();
         db.addLeague(league);
     }
 
-    public void addManager(Manager manager) {
+    public void addManager(Manager manager) throws AlreadyExistException {
+        if(db.existMember(manager.getUserMail()))
+            throw new AlreadyExistException();
         db.addManager(manager);
     }
 
-    public void addPlayer(Player player) {
+    public void addPlayer(Player player) throws AlreadyExistException {
+        if(db.existMember(player.getUserMail()))
+            throw new AlreadyExistException();
         db.addPlayer(player);
     }
 
-    public void addCoach(Coach coach) {
+    public void addCoach(Coach coach) throws AlreadyExistException {
+        if(db.existMember(coach.getUserMail()))
+            throw new AlreadyExistException();
         db.addCoach(coach);
     }
 
-    public void addOwner(Owner owner) {
+    public void addOwner(Owner owner) throws AlreadyExistException {
+        if(db.existMember(owner.getUserMail()))
+            throw new AlreadyExistException();
         db.addOwner(owner);
     }
 
-    public void addSystemManager(SystemManager systemManager) {
+    public void addSystemManager(SystemManager systemManager) throws AlreadyExistException {
+        if(db.existMember(systemManager.getUserMail()))
+            throw new AlreadyExistException();
         db.addSystemManager(systemManager);
     }
 
-    public void addTeam(Team team) {
+    public void addTeam(Team team) throws AlreadyExistException{
+        if(db.existTeam(team.getName()))
+            throw new AlreadyExistException();
         db.addTeam(team);
     }
 
     /***************************************remove function******************************************/
-    public void removeLeague(String leagueName) {
+    public void removeLeague(String leagueName) throws LeagueNotExist{
+        if(! db.existLeague(leagueName))
+            throw new LeagueNotExist();
         db.removeLeague(leagueName);
     }
 
-    public void removeSeason(String year) {
+    public void removeSeason(String year)  throws SeasonNotExist{
+        if(! db.existSeason(year))
+            throw new SeasonNotExist();
         db.removeSeason(year);
     }
 
-    public void removeTeam(String name) {
+    public void removeTeam(String name) throws TeamNotExist{
+        if(! db.existTeam(name))
+            throw new TeamNotExist();
         db.removeTeam(name);
     }
-
-    /***************************************exist function******************************************/
-
-
 }
